PGB: initialized variables KFORM,NRY#,LV[]
Found that un-initialized variables in PGB caused failures when
PGBX6SV.DAT was empty.  This file is IN6/FL1(6) in input list file
PGB.BAK.  When it is empty, the READ() calls leave these variables with
bad values.  This commit sets them all to 0, manually.

Perhaps -finit-local-zero would be better, but that caused problems
before when I tried it.